Megoldott feladatok

+++ src/main/java/model/Grade.java
@@ -4,13 +4,13 @@ import java.util.Date;
 import java.util.List;
 
 public class Grade {
-    private int teacherId;
-    private int subjectId;
-    private int studentId;
+    private long teacherId;
+    private long subjectId;
+    private long studentId;
 
     private List<Integer> grades;
 
-    public Grade(int teacherId, int subjectId, int studentId, List<Integer> grades) {
+    public Grade(long teacherId, long subjectId, long studentId, List<Integer> grades) {
         this.teacherId = teacherId;
         this.subjectId = subjectId;
         this.studentId = studentId;
